% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{cpp_register}
\alias{cpp_register}
\title{Generates wrappers for registered C++ functions}
\usage{
cpp_register(
  path = ".",
  quiet = !is_interactive(),
  extension = c(".cpp", ".cc")
)
}
\description{
Generates wrappers for registered C++ functions
}
