% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp20-package.R
\docType{package}
\name{cpp20-package}
\alias{cpp20}
\alias{cpp20-package}
\title{cpp20: A C++20 API For Interacting With R}
\description{
A header-only API designed for advanced R data manipulation within C++. cpp20 emphasises safety, modern C++ syntax and high-performance computing through SIMD vectorisation and parallelisation. Leveraging C++20 concepts specific to R, it enables seamless template programming with a highly efficient API.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/NicChr/cpp20/issues}
}

}
\author{
\strong{Maintainer}: Nick Christofides \email{nick.christofides.r@gmail.com} (\href{https://orcid.org/0000-0002-9743-7342}{ORCID})

}
\keyword{internal}
